Renamed a dscalarm package to internal. See openhab/static-code-analysis#100
