+++ addons/binding/org.openhab.binding.dscalarm/src/main/java/org/openhab/binding/dscalarm/handler/TCPServerBridgeHandler.java
@@ -20,7 +20,7 @@ import java.net.UnknownHostException;
 
 import org.eclipse.smarthome.core.thing.Bridge;
 import org.eclipse.smarthome.core.thing.ThingStatus;
-import org.openhab.binding.dscalarm.config.TCPServerBridgeConfiguration;
+import org.openhab.binding.dscalarm.internal.config.TCPServerBridgeConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
