+++ addons/binding/org.openhab.binding.dscalarm/src/main/java/org/openhab/binding/dscalarm/internal/config/DSCAlarmPanelConfiguration.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.dscalarm.config;
+package org.openhab.binding.dscalarm.internal.config;
 
 /**
  * Configuration class for the DSC Alarm Panel Thing.
