+++ addons/binding/org.openhab.binding.dscalarm/src/main/java/org/openhab/binding/dscalarm/internal/factory/DSCAlarmHandlerFactory.java
@@ -23,11 +23,6 @@ import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.openhab.binding.dscalarm.DSCAlarmBindingConstants;
-import org.openhab.binding.dscalarm.config.DSCAlarmPartitionConfiguration;
-import org.openhab.binding.dscalarm.config.DSCAlarmZoneConfiguration;
-import org.openhab.binding.dscalarm.config.EnvisalinkBridgeConfiguration;
-import org.openhab.binding.dscalarm.config.IT100BridgeConfiguration;
-import org.openhab.binding.dscalarm.config.TCPServerBridgeConfiguration;
 import org.openhab.binding.dscalarm.handler.DSCAlarmBaseBridgeHandler;
 import org.openhab.binding.dscalarm.handler.EnvisalinkBridgeHandler;
 import org.openhab.binding.dscalarm.handler.IT100BridgeHandler;
@@ -36,6 +31,11 @@ import org.openhab.binding.dscalarm.handler.PanelThingHandler;
 import org.openhab.binding.dscalarm.handler.PartitionThingHandler;
 import org.openhab.binding.dscalarm.handler.TCPServerBridgeHandler;
 import org.openhab.binding.dscalarm.handler.ZoneThingHandler;
+import org.openhab.binding.dscalarm.internal.config.DSCAlarmPartitionConfiguration;
+import org.openhab.binding.dscalarm.internal.config.DSCAlarmZoneConfiguration;
+import org.openhab.binding.dscalarm.internal.config.EnvisalinkBridgeConfiguration;
+import org.openhab.binding.dscalarm.internal.config.IT100BridgeConfiguration;
+import org.openhab.binding.dscalarm.internal.config.TCPServerBridgeConfiguration;
 import org.openhab.binding.dscalarm.internal.discovery.DSCAlarmDiscoveryService;
 import org.osgi.framework.ServiceRegistration;
 import org.slf4j.Logger;
